Add files via upload
Store the first element:- Store the first element of the array in a temporary variable. This element will be moved to the end of the array after the rotation.. , 
Shift elements:- Shift all elements one position to the left. This can be done by iterating through the array starting from the second element and copying each element to its previous position... ,
Place the stored element:-:Place the element stored in the temporary variable at the end of the array...etc